cleanup + modulated
clean up the code so that it is in modules instead of a large 2k+ line file
