+++ core/__init__.py
@@ -1,0 +1,22 @@
+"""Core modules for KickDropsMiner"""
+from .config import Config
+from .browser import CookieManager, make_chrome_driver
+from .api import (
+    kick_is_live_by_api,
+    fetch_drops_campaigns_and_progress,
+    fetch_live_streamers_by_category,
+    is_campaign_expired
+)
+from .worker import StreamWorker
+
+__all__ = [
+    'Config',
+    'CookieManager',
+    'make_chrome_driver',
+    'kick_is_live_by_api',
+    'fetch_drops_campaigns_and_progress',
+    'fetch_live_streamers_by_category',
+    'is_campaign_expired',
+    'StreamWorker'
+]
+
