+++ ui/__init__.py
@@ -1,0 +1,5 @@
+"""UI module for KickDropsMiner"""
+from .app import App
+
+__all__ = ['App']
+
