+++ utils/__init__.py
@@ -1,0 +1,4 @@
+"""Utility modules for KickDropsMiner"""
+from .helpers import *
+from .translations import *
+
